+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.2
 
 - Added config option "megaRayquazaAllowed" to control whether mega rayquaza can be used.
+- Fixed /givemegastone only being usable by players.
 
 # 1.0.1